Updating secrets screams if a file is missing
Previously this may not have been a loud and noisy failure; but I
believe we need to scream if it is not there.

[skip ci]

+++ script/update_secrets.sh
@@ -34,7 +34,13 @@ files_to_copy="
 
 for f in $files_to_copy; do
     echo "=-=-=-=-=-=-=-= copy $f"
-    cp $secret_repo/curate_nd/$f config/$f
+    if [ -f $secret_repo/curate_nd/$f ];
+    then
+        cp $secret_repo/curate_nd/$f config/$f
+    else
+        echo "Fatal Error: File $f does not exist in $secret_repo/curate_nd"
+        exit 1
+    fi
 done
 echo "=-=-=-=-=-=-=-= copy secret_token.rb"
 cp $secret_repo/curate_nd/secret_token.rb config/initializers/secret_token.rb
